Conflicts:
	plugins/plugins.pro
	qrgui/dialogs/preferencesDialog.cpp
	qrgui/dialogs/preferencesDialog.h
	qrgui/mainwindow/mainWindow.cpp
	qrgui/mainwindow/mainWindow.h
	qrgui/mainwindow/mainWindow.ui
	qrgui/mainwindow/projectManager/projectManager.cpp
	qrgui/mainwindow/projectManager/projectManager.h
	qrgui/qrgui.pro
	qrgui/qrgui.qrc
	qrgui/qrgui_ru.qm
	qrgui/qrgui_ru.ts
	qrgui/toolPluginInterface/pluginConfigurator.h
	qrgui/view/editorView.cpp
	qrgui/view/editorView.h
	qrkernel/settingsManager.cpp
	qrrepo/private/classes/object.cpp
	qrrepo/private/client.cpp
	qrrepo/private/folderCompressor.cpp
	qrrepo/private/repository.h
	qrrepo/private/serializer.cpp
	qrrepo/private/serializer.h
	qrrepo/qrrepo.pro
	qrrepo/repoApi.h
	qrrepo/repoControlInterface.h
	qrutils/qrutils.pro
	qrutils/qrutils_ru.qm
	qrutils/qrutils_ru.ts

+++ qrutils/mathUtils/gaussNoise.h
@@ -2,10 +2,9 @@
 
 #include <QtCore/QtGlobal>
 
-#include "../utilsDeclSpec.h"
+#include <qrutils/utilsDeclSpec.h>
 
-namespace mathUtils
-{
+namespace mathUtils {
 
 qreal const mu = 0.5;
 qreal const var = 0.083; // 1/12
@@ -32,7 +31,7 @@ public:
 
 private:
 	static const int defaultApproximationLevel = 12;
-	static const qreal defaultDispersion = 1.0;
+	static const int defaultDispersion = 1;
 
 	/// Body of function 'generate'. Uses for various realizations of 'generate'
 	qreal genBody(unsigned int const approximationLevel, qreal const variance) const;
